Refactor and execute validation on user type
