+++ src/App.tsx
@@ -1,11 +1,14 @@
 import Signup from "./components/Signup";
 import { FormContextProvider } from "./contexts/FormContext";
+import { InputErrorContextProvider } from "./contexts/InputErrorContext";
 
 function App() {
   return (
     <>
       <FormContextProvider>
-        <Signup />
+        <InputErrorContextProvider>
+          <Signup />
+        </InputErrorContextProvider>
       </FormContextProvider>
     </>
   );
